MDEV-14443: check_grant mask grant_table entries with the deny mask
Previous code assumed that if any combination of rights is acceptable,
that means that the presence of the grant_table entries themselves
signifies that access is granted. However this is not true if the
deny_mask masks all bits present in the grant_tables. This is now
corrected.

TODO: column level denies however are not evaluated.