MDEV-28609 refine gtid-strict-mode to ignore same server-id gtid from the past on semisync slave
To provide semisync master crash-recovery a same server-id transaction
was let into execution on the semisync slave when the strict gtid
mode (see MDEV-27760).
While gtids from the past are errored out by the existing strict mode
policy such gtids can naturally appear in in the circular setup which
stopped to work seeing the gtid out-of-order error.

This is fixed now to ignore in the gtid strict mode
gtids existing in the slave's binlog, similarly to the same-server-id
ignore policy.