MDEV-26546 SIGSEGV's in spider_db_connect on SHOW TABLE and spider_db_mbase::connect (and SIGSEGV's in check_vcol_forward_refs and inline_mysql_mutex_lock)
Not the SPIDER issue - happens to INSERT DELAYED.
the field::make_new_field didn't copy the LONG_UNIQUE_HASH_FIELD
flag to the new field. Though the Delayed_insert::get_local_table
copies the field->vcol_info for this field. Ad a result
the parse_vcol_defs doesn't create the expression for that column
so the field->vcol_info->expr is NULL. Which leads to crash.

Another problem with the USING HASH key is thst the
parse_vcol_defs modifies the table->keys content. Then the same
parse_vcol_defs is called on the table copy that has keys already modified.
Particularly the key->key_part is moved to the virtual keypart with
the HASH value, and the key->user_defined_key_parts is set to 1.
Fixed by 'refreshing' key definitions before the call.